+++ backend/app/api/nutritonix/client.py
@@ -1,0 +1,7 @@
+import datetime as dt
+from sqlmodel import Field, Relationship, SQLModel
+
+class NutritionixValues(SQLModel):   
+   
+    
+    created_at = dt.datetime.now(dt.timezone.utc)
