+++ backend/app/api/routes/users.py
@@ -15,15 +15,8 @@ from app.core.security import get_password_hash, verify_password
 from app.models import (
     Item,
     Message,
-    UpdatePassword,
-    User,
-    UserCreate,
-    UserPublic,
-    UserRegister,
-    UsersPublic,
-    UserUpdate,
-    UserUpdateMe,
 )
+from api.user.models import User, UserCreate,UserPublic,UserRegister, UsersPublic,UserUpdate,UserUpdateMe
 from app.utils import generate_new_account_email, send_email
 
 router = APIRouter()
